+++ src/main/java/net/nomUtiliser/potatoClicker/logic/data/Upgrade.java
@@ -1,6 +1,8 @@
 package net.nomUtiliser.potatoClicker.logic.data;
 
+import java.math.BigInteger;
+
 public class Upgrade {
     public String id;
-    public int quantity;
+    public BigInteger quantity;
 }
